More reorganization and cleanup, impl conversions
