+++ src/iconforge/image_cache.rs
@@ -1,14 +1,12 @@
-use crate::iconforge::universal_icon::{UniversalIcon, UniversalIconData};
+use super::universal_icon::{UniversalIcon, UniversalIconData};
 use dashmap::DashMap;
-use dmi::dirs::{ALL_DIRS, CARDINAL_DIRS};
-use dmi::icon::{dir_to_dmi_index, Icon};
-use dmi::{dirs::Dirs, icon::IconState};
+use dmi::{
+    dirs::{Dirs, ALL_DIRS, CARDINAL_DIRS},
+    icon::{dir_to_dmi_index, Icon, IconState},
+};
 use image::DynamicImage;
 use once_cell::sync::Lazy;
-use std::fs::File;
-use std::hash::BuildHasherDefault;
-use std::io::BufReader;
-use std::sync::Arc;
+use std::{fs::File, hash::BuildHasherDefault, io::BufReader, sync::Arc};
 use tracy_full::zone;
 use twox_hash::XxHash64;
 
@@ -25,143 +23,143 @@ pub fn image_cache_clear() {
     ICON_STATES.clear();
 }
 
-/// Takes an UniversalIcon, gets its DMI, then picks out a UniversalIconData for the IconState.
-/// If flatten is true, will output only one dir and frame (defaulting to SOUTH/1 if unscoped) regardless of the input uni_icon
-/// Returns with True if the UniversalIconData is pre-cached (and shouldn't have new transforms applied)
-pub fn universal_icon_to_images(
-    uni_icon: &UniversalIcon,
-    sprite_name: &String,
-    cached: bool,
-    must_be_cached: bool,
-    flatten: bool,
-) -> Result<(Arc<UniversalIconData>, bool), String> {
-    zone!("universal_icon_to_images");
-    if cached {
-        zone!("check_image_cache");
-        if let Some(entry) = ICON_STATES.get(uni_icon) {
-            return Ok((entry.value().to_owned(), true));
-        }
-        if must_be_cached {
-            return Err(format!("Image was requested but does not exist in the cache. It's likely that the icon state doesn't exist: {uni_icon} - while generating '{sprite_name}'"));
-        }
-    }
-    let dmi = filepath_to_dmi(&uni_icon.icon_file)?;
-    let mut matched_state: Option<&IconState> = None;
-    {
-        zone!("match_icon_state");
-        for icon_state in &dmi.states {
-            if icon_state.name == uni_icon.icon_state {
-                matched_state = Some(icon_state);
-                break;
+impl UniversalIcon {
+    /// Gets this icon's associated DMI, then picks out a UniversalIconData for the IconState.
+    /// If flatten is true, will output only one dir and frame (defaulting to SOUTH/1 if unscoped) regardless of the input uni_icon
+    /// Returns with True if the UniversalIconData is pre-cached (and shouldn't have new transforms applied)
+    pub fn get_image_data(
+        &self,
+        sprite_name: &String,
+        cached: bool,
+        must_be_cached: bool,
+        flatten: bool,
+    ) -> Result<(Arc<UniversalIconData>, bool), String> {
+        zone!("universal_icon_to_image_data");
+        if cached {
+            zone!("check_image_cache");
+            if let Some(entry) = ICON_STATES.get(self) {
+                return Ok((entry.value().to_owned(), true));
+            }
+            if must_be_cached {
+                return Err(format!("Image was requested but does not exist in the cache. It's likely that the icon state doesn't exist: {self} - while generating '{sprite_name}'"));
             }
         }
-    }
-    let state = match matched_state {
-        Some(state) => state,
-        None => {
-            return Err(format!(
-                "Could not find associated icon state {} for {sprite_name}",
-                uni_icon.icon_state
-            ));
+        let dmi = filepath_to_dmi(&self.icon_file)?;
+        let mut matched_state: Option<&IconState> = None;
+        {
+            zone!("match_icon_state");
+            for icon_state in &dmi.states {
+                if icon_state.name == self.icon_state {
+                    matched_state = Some(icon_state);
+                    break;
+                }
+            }
         }
-    };
+        let state = match matched_state {
+            Some(state) => state,
+            None => {
+                return Err(format!(
+                    "Could not find associated icon state {} for {sprite_name}",
+                    self.icon_state
+                ));
+            }
+        };
 
-    let mut dirs = state.dirs as usize;
-    let mut dir_index = 0;
+        let mut dirs = state.dirs as usize;
+        let mut dir_index = 0;
 
-    if let Some(dir_bits) = uni_icon.dir {
-        // Consider 0 to be "unscoped"
-        if dir_bits > 0 {
-            dirs = 1;
-            dir_index = match Dirs::from_bits(dir_bits) {
-                Some(dir) => {
-                    if (state.dirs == 1 && dir != Dirs::SOUTH)
-                        || (state.dirs == 4 && !CARDINAL_DIRS.contains(&dir))
-                        || (state.dirs == 8 && !ALL_DIRS.contains(&dir))
-                    {
-                        return Err(format!(
-                            "Dir specified {dir} is not in the set of valid dirs ({} dirs) for icon_state \"{}\" for {sprite_name}", state.dirs, state.name
-                        ));
-                    }
-                    match dir_to_dmi_index(&dir) {
-                        Some(index) => index,
-                        None => {
+        if let Some(dir_bits) = self.dir {
+            // Consider 0 to be "unscoped"
+            if dir_bits > 0 {
+                dirs = 1;
+                dir_index = match Dirs::from_bits(dir_bits) {
+                    Some(dir) => {
+                        if (state.dirs == 1 && dir != Dirs::SOUTH)
+                            || (state.dirs == 4 && !CARDINAL_DIRS.contains(&dir))
+                            || (state.dirs == 8 && !ALL_DIRS.contains(&dir))
+                        {
                             return Err(format!(
-                                "Invalid dir in dir ordering {dir} for {sprite_name}"
+                                "Dir specified {dir} is not in the set of valid dirs ({} dirs) for icon_state \"{}\" for {sprite_name}", state.dirs, state.name
                             ));
                         }
+                        match dir_to_dmi_index(&dir) {
+                            Some(index) => index,
+                            None => {
+                                return Err(format!(
+                                    "Invalid dir in dir ordering {dir} for {sprite_name}"
+                                ));
+                            }
+                        }
                     }
-                }
-                None => {
-                    return Err(format!("Invalid dir number {dir_bits} for {sprite_name}"));
-                }
-            };
+                    None => {
+                        return Err(format!("Invalid dir number {dir_bits} for {sprite_name}"));
+                    }
+                };
+            } else if flatten {
+                dirs = 1;
+                dir_index = 0;
+            }
         } else if flatten {
             dirs = 1;
             dir_index = 0;
         }
-    } else if flatten {
-        dirs = 1;
-        dir_index = 0;
-    }
 
-    let mut frames = state.frames as usize;
-    let mut frame_offset: usize = 0;
+        let mut frames = state.frames as usize;
+        let mut frame_offset: usize = 0;
 
-    if let Some(frame) = uni_icon.frame {
-        // Consider 0 to be "unscoped"
-        // Also no underflow please
-        if frame > 0 {
-            frames = 1;
-            frame_offset = frame as usize - 1;
-            if state.frames < frame {
-                return Err(format!(
-                    "Specified frame \"{frame}\" is larger than the number of frames ({}) for icon_state \"{}\" in sprite \"{sprite_name}\"",
-                    state.frames, state.name
-                ));
+        if let Some(frame) = self.frame {
+            // Consider 0 to be "unscoped"
+            // Also no underflow please
+            if frame > 0 {
+                frames = 1;
+                frame_offset = frame as usize - 1;
+                if state.frames < frame {
+                    return Err(format!(
+                        "Specified frame \"{frame}\" is larger than the number of frames ({}) for icon_state \"{}\" in sprite \"{sprite_name}\"",
+                        state.frames, state.name
+                    ));
+                }
+            } else if flatten {
+                frames = 1;
+                frame_offset = 0;
             }
         } else if flatten {
             frames = 1;
             frame_offset = 0;
         }
-    } else if flatten {
-        frames = 1;
-        frame_offset = 0;
-    }
 
-    let mut images: Vec<DynamicImage> = Vec::new();
+        let mut images: Vec<DynamicImage> = Vec::new();
 
-    for frame_index in frame_offset..(frame_offset + frames) {
-        for dir_offset in dir_index..(dir_index + dirs) {
-            match state
-                .images
-                .get((frame_index * state.dirs as usize) + dir_offset)
-            {
-                Some(image) => images.push(image.clone()),
-                None => {
-                    return Err(format!("Somehow got out of bounds image for dir {dir_index} and frame {frame_offset} on {sprite_name}!"));
+        for frame_index in frame_offset..(frame_offset + frames) {
+            for dir_offset in dir_index..(dir_index + dirs) {
+                match state
+                    .images
+                    .get((frame_index * state.dirs as usize) + dir_offset)
+                {
+                    Some(image) => images.push(image.clone()),
+                    None => {
+                        return Err(format!("Somehow got out of bounds image for dir {dir_index} and frame {frame_offset} on {sprite_name}!"));
+                    }
                 }
             }
         }
+
+        let result = Arc::new(UniversalIconData {
+            images,
+            frames: frames as u32,
+            dirs: dirs as u8,
+            delay: if frames > 1 {
+                state.delay.to_owned()
+            } else {
+                None
+            },
+            loop_flag: state.loop_flag,
+            rewind: state.rewind,
+        });
+
+        // Don't insert into the cache here, because the cache should only contain transformed images.
+        Ok((result, false))
     }
-
-    let result = Arc::new(UniversalIconData {
-        images,
-        frames: frames as u32,
-        dirs: dirs as u8,
-        delay: if frames > 1 {
-            state.delay.to_owned()
-        } else {
-            None
-        },
-        loop_flag: state.loop_flag,
-        rewind: state.rewind,
-    });
-
-    // Don't insert, because the cache should only contain transformed images.
-    //ICON_STATES.insert(uni_icon.to_owned(), result.to_owned());
-
-    Ok((result, false))
 }
 
 pub fn cache_transformed_images(uni_icon: &UniversalIcon, image_data: Arc<UniversalIconData>) {
